third_party/blink/web_tests/fast/events/touch/compositor-touch-hit-rects-many-expected.txt

Tests that there is an upper limit on the number of hit rects generated per layer. http://crbug.com/299177.

Test node has 99 children with 100 rects
Test node has 100 children with 101 rects
FAIL - got 101 rects, expected 1.
Test node has 101 children with 102 rects
FAIL - got 102 rects, expected 1.
manychildren: layer(308x1228) has hit test rect (0,0 290x12)
manychildren: layer(308x1228) has hit test rect (301,21 7x7)
manychildren: layer(308x1228) has hit test rect (301,33 7x7)
manychildren: layer(308x1228) has hit test rect (301,45 7x7)
manychildren: layer(308x1228) has hit test rect (301,57 7x7)
manychildren: layer(308x1228) has hit test rect (301,69 7x7)
manychildren: layer(308x1228) has hit test rect (301,81 7x7)
manychildren: layer(308x1228) has hit test rect (301,93 7x7)
manychildren: layer(308x1228) has hit test rect (301,105 7x7)
manychildren: layer(308x1228) has hit test rect (301,117 7x7)
manychildren: layer(308x1228) has hit test rect (301,129 7x7)
manychildren: layer(308x1228) has hit test rect (301,141 7x7)
manychildren: layer(308x1228) has hit test rect (301,153 7x7)
manychildren: layer(308x1228) has hit test rect (301,165 7x7)
manychildren: layer(308x1228) has hit test rect (301,177 7x7)
manychildren: layer(308x1228) has hit test rect (301,189 7x7)
manychildren: layer(308x1228) has hit test rect (301,201 7x7)
manychildren: layer(308x1228) has hit test rect (301,213 7x7)
manychildren: layer(308x1228) has hit test rect (301,225 7x7)
manychildren: layer(308x1228) has hit test rect (301,237 7x7)
manychildren: layer(308x1228) has hit test rect (301,249 7x7)
manychildren: layer(308x1228) has hit test rect (301,261 7x7)
manychildren: layer(308x1228) has hit test rect (301,273 7x7)
manychildren: layer(308x1228) has hit test rect (301,285 7x7)
manychildren: layer(308x1228) has hit test rect (301,297 7x7)
manychildren: layer(308x1228) has hit test rect (301,309 7x7)
manychildren: layer(308x1228) has hit test rect (301,321 7x7)
manychildren: layer(308x1228) has hit test rect (301,333 7x7)
manychildren: layer(308x1228) has hit test rect (301,345 7x7)
manychildren: layer(308x1228) has hit test rect (301,357 7x7)
manychildren: layer(308x1228) has hit test rect (301,369 7x7)
manychildren: layer(308x1228) has hit test rect (301,381 7x7)
manychildren: layer(308x1228) has hit test rect (301,393 7x7)
manychildren: layer(308x1228) has hit test rect (301,405 7x7)
manychildren: layer(308x1228) has hit test rect (301,417 7x7)
manychildren: layer(308x1228) has hit test rect (301,429 7x7)
manychildren: layer(308x1228) has hit test rect (301,441 7x7)
manychildren: layer(308x1228) has hit test rect (301,453 7x7)
manychildren: layer(308x1228) has hit test rect (301,465 7x7)
manychildren: layer(308x1228) has hit test rect (301,477 7x7)
manychildren: layer(308x1228) has hit test rect (301,489 7x7)
manychildren: layer(308x1228) has hit test rect (301,501 7x7)
manychildren: layer(308x1228) has hit test rect (301,513 7x7)
manychildren: layer(308x1228) has hit test rect (301,525 7x7)
manychildren: layer(308x1228) has hit test rect (301,537 7x7)
manychildren: layer(308x1228) has hit test rect (301,549 7x7)
manychildren: layer(308x1228) has hit test rect (301,561 7x7)
manychildren: layer(308x1228) has hit test rect (301,573 7x7)
manychildren: layer(308x1228) has hit test rect (301,585 7x7)
manychildren: layer(308x1228) has hit test rect (301,597 7x7)
manychildren: layer(308x1228) has hit test rect (301,609 7x7)
manychildren: layer(308x1228) has hit test rect (301,621 7x7)
manychildren: layer(308x1228) has hit test rect (301,633 7x7)
manychildren: layer(308x1228) has hit test rect (301,645 7x7)
manychildren: layer(308x1228) has hit test rect (301,657 7x7)
manychildren: layer(308x1228) has hit test rect (301,669 7x7)
manychildren: layer(308x1228) has hit test rect (301,681 7x7)
manychildren: layer(308x1228) has hit test rect (301,693 7x7)
manychildren: layer(308x1228) has hit test rect (301,705 7x7)
manychildren: layer(308x1228) has hit test rect (301,717 7x7)
manychildren: layer(308x1228) has hit test rect (301,729 7x7)
manychildren: layer(308x1228) has hit test rect (301,741 7x7)
manychildren: layer(308x1228) has hit test rect (301,753 7x7)
manychildren: layer(308x1228) has hit test rect (301,765 7x7)
manychildren: layer(308x1228) has hit test rect (301,777 7x7)
manychildren: layer(308x1228) has hit test rect (301,789 7x7)
manychildren: layer(308x1228) has hit test rect (301,801 7x7)
manychildren: layer(308x1228) has hit test rect (301,813 7x7)
manychildren: layer(308x1228) has hit test rect (301,825 7x7)
manychildren: layer(308x1228) has hit test rect (301,837 7x7)
manychildren: layer(308x1228) has hit test rect (301,849 7x7)
manychildren: layer(308x1228) has hit test rect (301,861 7x7)
manychildren: layer(308x1228) has hit test rect (301,873 7x7)
manychildren: layer(308x1228) has hit test rect (301,885 7x7)
manychildren: layer(308x1228) has hit test rect (301,897 7x7)
manychildren: layer(308x1228) has hit test rect (301,909 7x7)
manychildren: layer(308x1228) has hit test rect (301,921 7x7)
manychildren: layer(308x1228) has hit test rect (301,933 7x7)
manychildren: layer(308x1228) has hit test rect (301,945 7x7)
manychildren: layer(308x1228) has hit test rect (301,957 7x7)
manychildren: layer(308x1228) has hit test rect (301,969 7x7)
manychildren: layer(308x1228) has hit test rect (301,981 7x7)
manychildren: layer(308x1228) has hit test rect (301,993 7x7)
manychildren: layer(308x1228) has hit test rect (301,1005 7x7)
manychildren: layer(308x1228) has hit test rect (301,1017 7x7)
manychildren: layer(308x1228) has hit test rect (301,1029 7x7)
manychildren: layer(308x1228) has hit test rect (301,1041 7x7)
manychildren: layer(308x1228) has hit test rect (301,1053 7x7)
manychildren: layer(308x1228) has hit test rect (301,1065 7x7)
manychildren: layer(308x1228) has hit test rect (301,1077 7x7)
manychildren: layer(308x1228) has hit test rect (301,1089 7x7)
manychildren: layer(308x1228) has hit test rect (301,1101 7x7)
manychildren: layer(308x1228) has hit test rect (301,1113 7x7)
manychildren: layer(308x1228) has hit test rect (301,1125 7x7)
manychildren: layer(308x1228) has hit test rect (301,1137 7x7)
manychildren: layer(308x1228) has hit test rect (301,1149 7x7)
manychildren: layer(308x1228) has hit test rect (301,1161 7x7)
manychildren: layer(308x1228) has hit test rect (301,1173 7x7)
manychildren: layer(308x1228) has hit test rect (301,1185 7x7)
manychildren: layer(308x1228) has hit test rect (301,1197 7x7)
manychildren: layer(308x1228) has hit test rect (301,1209 7x7)
manychildren: layer(308x1228) has hit test rect (301,1221 7x7)
